requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.1
dataclasses-json==0.6.1
exceptiongroup==1.1.3
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.10.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.321
langsmith==0.0.49
llvmlite==0.41.1
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2
numba==0.58.1
numpy==1.26.1
openai==0.28.1
openai-whisper==20230918
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
sympy==1.12
tenacity==8.2.3
tiktoken==0.3.3
torch==2.1.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
zipp==3.17.0
